requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==0.7.4
Markdown==3.7
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
